# academy_clean/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for أكاديمية البرمجة المتكاملة
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course with lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
...
  /courses/{courseId}/lessons:
    post:
      operationId: createLesson
      tags: [courses]
      summary: Add a lesson to a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

  # ─── Reviews ─────────────────────────────────────────────────────────────────
  /courses/{courseId}/reviews:
    get:
      operationId: getCourseReviews
      tags: [reviews]
      summary: Get reviews for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── Repositories ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        language:
          type: string
        creatorId:
          type: string
        creatorName:
          type: string
        creatorAvatar:
          type: string
        averageRating:
          type: number
        totalReviews:
          type: integer
        totalLessons:
          type: integer
        totalEnrollments:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, category, level, creatorId, creatorName, averageRating, totalReviews, totalLessons, createdAt]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
          required: [lessons]
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
      required: [courses, total]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        language:
          type: string
          default: ar
        creatorId:
          type: string
      required: [title, description, category, level, creatorId]
    Lesson:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        duration:
          type: integer
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, courseId, title, order, createdAt]
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        duration:
          type: integer
        order:
          type: integer
      required: [title, order]

    # ─── Review schemas ───────────────────────────────────────────────────────
    Review:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        userId:
          type: string
        userName:
          type: string
        userAvatar:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, courseId, userId, userName, rating, createdAt]
    CreateReviewBody:
      type: object
      properties:
        userId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [userId, rating]

    # ─── Repository schemas ───────────────────────────────────────────────────
...